Rework integer type selection
Following the recent gosec-related int work (to annotate known-safe
conversions), this furthers that by:

* using ints instead of uints for healthcheck configuration and
  tracking
* using uint32 instead of int for IP pool sizes, since their use is
  tied to uint32 representations of IP addresses
* using fmt.Sprintf instead of manual string concatenation to
  construct the connectivity verification command
* using int32 for nat discovery IPPortPair

+++ pkg/natdiscovery/request_send.go
@@ -87,11 +87,11 @@ func (nd *natDiscovery) sendCheckRequestToTargetIP(remoteNAT *remoteEndpointNAT,
 		},
 		UsingSrc: &natproto.IPPortPair{
 			IP:   sourceIP,
-			Port: uint32(nd.serverPort),
+			Port: nd.serverPort,
 		},
 		UsingDst: &natproto.IPPortPair{
 			IP:   targetIP,
-			Port: uint32(targetPort),
+			Port: targetPort,
 		},
 	}
